Move osu simulation to subcommand
